Move NewWorkspace page to SharedScreens
