+++ src/libs/Navigation/AppNavigator/AuthScreens.js
@@ -62,11 +62,9 @@ import Timers from '../../Timers';
 import WorkspaceSettingsDrawerNavigator from './WorkspaceSettingsDrawerNavigator';
 import spacing from '../../../styles/utilities/spacing';
 import CardOverlay from '../../../components/CardOverlay';
-import defaultScreenOptions from './defaultScreenOptions';
 import * as API from '../../API';
 import {setLocale} from '../../actions/App';
 import {cleanupSession} from '../../actions/Session';
-import WorkspaceNew from '../../../pages/workspace/WorkspaceNew';
 
 Onyx.connect({
     key: ONYXKEYS.MY_PERSONAL_DETAILS,
@@ -308,11 +306,6 @@ class AuthScreens extends React.Component {
                         },
                     }}
                     component={MainDrawerNavigator}
-                />
-                <RootStack.Screen
-                    name="WorkspaceNew"
-                    options={defaultScreenOptions}
-                    component={WorkspaceNew}
                 />
 
                 {/* These are the various modal routes */}
