+++ src/pages/Transition/TransitionNewWorkspace.js
@@ -2,7 +2,7 @@ import React, {Component} from 'react';
 import * as Policy from '../../libs/actions/Policy';
 import FullScreenLoadingIndicator from '../../components/FullscreenLoadingIndicator';
 
-class WorkspaceNew extends Component {
+class TransitionNewWorkspace extends Component {
     componentDidMount() {
         // After the workspace is created, the user will automatically be directed to its settings page
         Policy.create();
@@ -13,4 +13,4 @@ class WorkspaceNew extends Component {
     }
 }
 
-export default WorkspaceNew;
+export default TransitionNewWorkspace;
